Godot metronome using timers
I have been experimenting with a turn-based rpg/rhythm game idea and so far this is the model I use to do everything that is related to a BPM.  I am posting because I want to seek feedback from the community on how to improve.  So far my metronome seems to work but I have not implemented anything in the process function or any flashy graphics.
Cheers